requirements.txt
# Core dependencies
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Web UI
gradio>=4.0.0

# Video processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Deep learning (optional - for advanced removal)
torch>=2.0.0
torchvision>=0.15.0

# Utilities
tqdm>=4.65.0
loguru>=0.7.0
pyyaml>=6.0
